backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.121.2",
    "passlib>=1.7.4",
    "pydantic-settings>=2.12.0",
    "pydantic[email]>=2.12.4",
    "pymongo>=4.15.4",
    "python-jose[cryptography]>=3.5.0",
    "requests>=2.32.5",
    "sentry-sdk>=2.44.0",
    "uvicorn[standard]>=0.38.0",
]
